updated gallery to responsive grid and removed mobile gallery
